Next iteration for a productive "custom" CleverReach Channel
This upload function is able to use tags instead of starting a broadcast. With an THEA automation in cleverreach you are able to sendout emails in near-real-time (max. 5 minutes in cleverreach). Tags can be defined as comma separated list like `test.tag, shop.newcustomer` and can be influenced with a `-` prefix or `*`.